Add files via upload
To ensure the proper functioning of the module, the connecthandlerc9800 module available in the python-wifi-modules repository, as well as the following libraries, are required:
- json
- time
- textfsm

In addition to the module and libraries, templates appropriate for the commands we issue are needed in order to transform the output data into readable and useful data. Some templates can be found in the following repository of mine: templates_TextFSM_wireless_Cisco.